pyproject.toml
[project]
name = "repl-nix-recipemanagerassistant"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "psycopg2-binary>=2.9.10",
    "streamlit>=1.39.0",
]
